backend/requirements.txt
# tensorflow==2.10.0
# keras==2.10.0
# padasip==1.2.2
# scikit-learn==1.5.1
pandas
numpy
plotly==5.24.1
python-dotenv==0.21.0
supabase==2.15.0
# lightgbm==4.6.0
fastapi
uvicorn
pydantic